consolidate bspline code
