--- pydl/pydlutils/bspline/__init__.py
@@ -1,9 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the bspline directory of idlutils.
-"""
-from .bspline import bspline
-from .cholesky_band import cholesky_band
-from .cholesky_solve import cholesky_solve
-from .iterfit import iterfit
