--- pydl/pydlutils/bspline/tests/__init__.py
@@ -1,5 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This is the pydl/pydlutils/bspline/tests directory.
-"""
